Restructures file architecture
